move MasterSlaveLoadBalanceAlgorithm to master-slave-core-api module
