+++ master-slave-core/master-slave-core-common/src/main/java/org/apache/shardingsphere/masterslave/core/rule/MasterSlaveDataSourceRule.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.core.rule;
+package org.apache.shardingsphere.masterslave.core.rule;
 
 import lombok.AccessLevel;
 import lombok.Getter;
