+++ master-slave-core/master-slave-core-common/src/main/resources/META-INF/services/org.apache.shardingsphere.underlying.common.rule.ShardingSphereRuleBuilder
@@ -15,5 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.core.rule.ShardingRuleBuilder
-org.apache.shardingsphere.core.rule.ShadowRuleBuilder
+org.apache.shardingsphere.masterslave.core.rule.MasterSlaveRuleBuilder
