+++ sharding-core/sharding-core-common/src/test/java/org/apache/shardingsphere/core/rule/MasterSlaveDataSourceRuleTest.java
@@ -19,6 +19,7 @@ package org.apache.shardingsphere.core.rule;
 
 import org.apache.shardingsphere.masterslave.api.config.LoadBalanceStrategyConfiguration;
 import org.apache.shardingsphere.masterslave.api.config.MasterSlaveDataSourceConfiguration;
+import org.apache.shardingsphere.masterslave.core.rule.MasterSlaveDataSourceRule;
 import org.junit.Test;
 
 import java.util.Arrays;
